Bonds

Description

Documentation

data Bond

Purely value-level type (Int not Nat) representing a bond between two Elements and the order of said bond

Constructors

Bond Int Element Element

bondLength :: Floating a => Bond -> a

Yields the length of some Bond approximated by the sum of covalent radii

reflexBondE :: Floating a => Element -> a

Very rough approximation of "reflexive" single bond energy (A-A) Regression on covalent radius, valence electrons, and effective nuclear charge mirroring Coulomb's law

kJ2eV :: Floating a => a -> a

Helper function, eV ~= kJ/96.485

eV2kJ :: Floating a => a -> a

Helper function, kJ ~= eV*96.485

kJ2kCal :: Floating a => a -> a

Helper function, kCal ~= kJ/4.184

kCal2kJ :: Floating a => a -> a

Helper function, kJ ~= kCal*4.184

singBondE :: Floating a => Element -> Element -> a

Single bond energy based on Pauling electronegativity difference for dissociation energy

Ed(AB) = 0.5(Ed(AA) + Ed(BB)) + (XA - XB)^2

Either very close or not at all

bondEnergy :: Floating a => Bond -> a

Higher order bond energies, small dataset used for a regression between single bond energy and bond length

totalBondEnergy :: (Molecular m, Floating a) => m -> a

Sum of bondEnergy/ies within some Molecular compound

shatterToBonds :: Molecule -> [Bond]

Break apart some Molecule into component Bonds