Elements

Description

Documentation

data Element

All elements Hydrogen through Oganesson

Constructors

H

He

Li

Be

B

C

N

O

F

Ne

Na

Mg

Al

Si

P

S

Cl

Ar

K

Ca

Sc

Ti

V

Cr

Mn

Fe

Co

Ni

Cu

Zn

Ga

Ge

As

Se

Br

Kr

Rb

Sr

Y

Zr

Nb

Mo

Tc

Ru

Rh

Pd

Ag

Cd

In

Sn

Sb

Te

I

Xe

Cs

Ba

La

Ce

Pr

Nd

Pm

Sm

Eu

Gd

Tb

Dy

Ho

Er

Tm

Yb

Lu

Hf

Ta

W

Re

Os

Ir

Pt

Au

Hg

Tl

Pb

Bi

Po

At

Rn

Fr

Ra

Ac

Th

Pa

U

Np

Pu

Am

Cm

Bk

Cf

Es

Fm

Md

No

Lr

Rf

Db

Sg

Bh

Hs

Mt

Ds

Rg

Cn

Nh

Fl

Mc

Lv

Ts

Og

class KnownElem (a :: Element)

Type class representing a reifiable Element

Minimal complete definition

elemSing

type KnownCharge (st :: Charge) = KnownNat (NaturalCharge st)

Derivation of reifiable Nat to reifiable Charge

type family ValenceE (e :: Element) :: Nat where ...

Type family representing the valence electrons (s+p electrons) for all representative elements All transition elements are assigned 2 for their s sublevel

Equations

ValenceE H = 1

ValenceE Li = 1

ValenceE Na = 1

ValenceE K = 1

ValenceE Rb = 1

ValenceE Cs = 1

ValenceE Fr = 1

ValenceE B = 3

ValenceE Al = 3

ValenceE In = 3

ValenceE Tl = 3

ValenceE Nh = 3

ValenceE C = 4

ValenceE Si = 4

ValenceE Ge = 4

ValenceE Sn = 4

ValenceE Pb = 4

ValenceE Fl = 4

ValenceE N = 5

ValenceE P = 5

ValenceE As = 5

ValenceE Sb = 5

ValenceE Bi = 5

ValenceE Mc = 5

ValenceE O = 6

ValenceE S = 6

ValenceE Se = 6

ValenceE Te = 6

ValenceE Po = 6

ValenceE Lv = 6

ValenceE F = 7

ValenceE Cl = 7

ValenceE Br = 7

ValenceE I = 7

ValenceE At = 7

ValenceE Ts = 7

ValenceE Ne = 8

ValenceE Ar = 8

ValenceE Kr = 8

ValenceE Xe = 8

ValenceE Rn = 8

ValenceE Og = 8

ValenceE _ = 2

valenceE :: Integral a => Element -> a

Value-level ValenceE

elemVal :: forall a. KnownElem a => Proxy a -> Element

Reify Element with Proxy

elemValI :: forall a. KnownElem a => Element

Reify Element independent of Proxy

toAtomic :: Element -> Int

Convert Element to 1-indexed atomic number

fromAtomic :: Int -> Element

Convert atomic number to Element

data Charge

Type representing a charge using Naturals. Useful for constraints at type level

Constructors

Neutral

Pos Nat

Neg Nat

charge :: Integral a => Charge -> a

Converts a Charge value to a signed Integral value

type family AddChg (c1 :: Charge) (c2 :: Charge) :: Charge where ...

Type family summing two Charges, identical to (+) in Num instance

Equations

AddChg Neutral c2 = c2

AddChg c1 Neutral = c1

AddChg (Neg a) (Neg b) = Neg (a + b)

AddChg (Neg a) (Pos b) = If (a == b) Neutral (If (a <=? b) (Pos (b - a)) (Neg (a - b)))

AddChg (Pos a) (Neg b) = AddChg (Neg a) (Pos b)

type family MulChg (n :: Nat) (c :: Charge) :: Charge where ...

Type family yielding some Charge multipled by a Natural. Both less versatile and less verbose than straight Charge-Charge multiplication

Equations

MulChg _ Neutral = Neutral

MulChg n (Pos c) = Pos (n * c)

MulChg n (Neg c) = Neg (n * c)

type family NaturalCharge (s :: Charge) = (n :: Nat) | n -> s where ...

Converts a Charge to an unwrapped Natural of value Charge + 3

Equations

NaturalCharge (Neg 3) = 0

NaturalCharge (Neg 2) = 1

NaturalCharge (Neg 1) = 2

NaturalCharge Neutral = 3

NaturalCharge (Pos 1) = 4

NaturalCharge (Pos 2) = 5

NaturalCharge (Pos 3) = 6

data Species (state :: Charge) tp

Type representing a species (i.e. in solution) Can be monatomic or binary polyatomic Polyatomic Ions are represented as two sets of Atoms

data Atoms

Type representing a number of atoms of a given Element

Constructors

Atoms Element Int

mkAtom :: forall a. KnownElem a => Species Neutral Element

Reify an Element as a Neutral Species

asSpecies :: t -> Species Neutral t

Convert an Element value to Neutral Species

type IonRep (e :: Element) = Species (VeToIon (ValenceE e)) Element

Useful type synonym for the (likely) common ion of a representative Element

mkIonRep :: forall a. KnownElem a => IonRep a

Reify a representative Element as its common ion

mkIonPoly :: forall e1 ct1 e2 ct2 charge. (KnownElem e1, KnownNat ct1, KnownElem e2, KnownNat ct2, ValidPolar2 e1 ct1 e2 ct2 charge) => Species charge (Atoms, Atoms)

Reify a binary polyatomic ion For example, Nitrate:

mkIonPoly @N @1 @O @3 @(Neg 1)

specValue :: Species a t -> t

Returns the value wrapped by a Species

specCharge :: forall a c n t. (KnownNat c, c ~ NaturalCharge a, Integral n) => Species a t -> n

Returns the charge of a value-level Species

chargeValI :: forall st n. (KnownNat n, n ~ NaturalCharge st) => Charge

Reifies a Charge independent of Proxy

type InvalidOxidationState (e :: Element) (target :: Nat) = ((Text "Invalid Ox State for " :<>: ShowType e) :<>: Text ": ") :<>: ShowType target

Useful type error for when an invalid oxidation state is presented

type family OctetExpandable (e :: Element) (target :: Nat) :: Constraint where ...

Constraint representing an Element with a valid number of bonded electrons

Equations

OctetExpandable Cl target = target <= 7

OctetExpandable N target = target <= 5

OctetExpandable e target = If (IsOctetExpandable (ValenceE e) target) (() :: Constraint) (TypeError (InvalidOxidationState e target))

type IsOctetExpandable (valence :: Nat) (target :: Nat) = (valence == target) || ((valence == (target + 2)) || ((valence == (target + 4)) || (valence == (target + 6))))

Boolean form of OctetExpandable accepting a raw number of valence electrons. Assumes representative and standard case

isOctetExpandable :: Integral a => a -> a -> Bool

Value-level IsOctetExpandable

type family ValidPolar2 (e1 :: Element) (ct1 :: Nat) (e2 :: Element) (ct2 :: Nat) (charge :: Charge) :: Constraint where ...

Constraint representing the a valid binary compound consisting of polar bonds between two representative Elements given their likely oxidation states and the final Charge of the compound. Oxygen, Nitrogen, and Chlorine are given special consideration. For all others, free electrons and both electrons of a split pair may bond

Equations

ValidPolar2 e1 ct1 O ct2 Neutral = If (IsOctetExpandable (ValenceE e1) (Div (ct2 * 2) ct1) || (IsOctetExpandable (ValenceE e1) (Div ct2 ct1) || IsOctetExpandable (ValenceE e1) (Div ct2 (2 * ct1)))) (() :: Constraint) (TypeError ((((((Text "No Valid Ox State for " :<>: ShowType e1) :<>: Text " in ") :<>: ShowType e1) :<>: ShowType ct1) :<>: Text "O") :<>: ShowType ct2))

ValidPolar2 e1 ct1 e2 ct2 Neutral = OctetExpandable e1 (Div (ct2 * (8 - ValenceE e2)) ct1)

ValidPolar2 e1 ct1 e2 ct2 (Pos charge) = OctetExpandable e1 (charge + Div (ct2 * (8 - ValenceE e2)) ct1)

ValidPolar2 e1 ct1 e2 ct2 (Neg charge) = OctetExpandable e1 (Div (ct2 * (8 - ValenceE e2)) ct1 - charge)

type family ToAtomic (e :: Element) = n | n -> e where ...

Type-level toAtomic, injective

Equations

ToAtomic H = 1

ToAtomic He = 2

ToAtomic Li = 3

ToAtomic Be = 4

ToAtomic B = 5

ToAtomic C = 6

ToAtomic N = 7

ToAtomic O = 8

ToAtomic F = 9

ToAtomic Ne = 10

ToAtomic Na = 11

ToAtomic Mg = 12

ToAtomic Al = 13

ToAtomic Si = 14

ToAtomic P = 15

ToAtomic S = 16

ToAtomic Cl = 17

ToAtomic Ar = 18

ToAtomic K = 19

ToAtomic Ca = 20

ToAtomic Sc = 21

ToAtomic Ti = 22

ToAtomic V = 23

ToAtomic Cr = 24

ToAtomic Mn = 25

ToAtomic Fe = 26

ToAtomic Co = 27

ToAtomic Ni = 28

ToAtomic Cu = 29

ToAtomic Zn = 30

ToAtomic Ga = 31

ToAtomic Ge = 32

ToAtomic As = 33

ToAtomic Se = 34

ToAtomic Br = 35

ToAtomic Kr = 36

ToAtomic Rb = 37

ToAtomic Sr = 38

ToAtomic Y = 39

ToAtomic Zr = 40

ToAtomic Nb = 41

ToAtomic Mo = 42

ToAtomic Tc = 43

ToAtomic Ru = 44

ToAtomic Rh = 45

ToAtomic Pd = 46

ToAtomic Ag = 47

ToAtomic Cd = 48

ToAtomic In = 49

ToAtomic Sn = 50

ToAtomic Sb = 51

ToAtomic Te = 52

ToAtomic I = 53

ToAtomic Xe = 54

ToAtomic Cs = 55

ToAtomic Ba = 56

ToAtomic La = 57

ToAtomic Ce = 58

ToAtomic Pr = 59

ToAtomic Nd = 60

ToAtomic Pm = 61

ToAtomic Sm = 62

ToAtomic Eu = 63

ToAtomic Gd = 64

ToAtomic Tb = 65

ToAtomic Dy = 66

ToAtomic Ho = 67

ToAtomic Er = 68

ToAtomic Tm = 69

ToAtomic Yb = 70

ToAtomic Lu = 71

ToAtomic Hf = 72

ToAtomic Ta = 73

ToAtomic W = 74

ToAtomic Re = 75

ToAtomic Os = 76

ToAtomic Ir = 77

ToAtomic Pt = 78

ToAtomic Au = 79

ToAtomic Hg = 80

ToAtomic Tl = 81

ToAtomic Pb = 82

ToAtomic Bi = 83

ToAtomic Po = 84

ToAtomic At = 85

ToAtomic Rn = 86

ToAtomic Fr = 87

ToAtomic Ra = 88

ToAtomic Ac = 89

ToAtomic Th = 90

ToAtomic Pa = 91

ToAtomic U = 92

ToAtomic Np = 93

ToAtomic Pu = 94

ToAtomic Am = 95

ToAtomic Cm = 96

ToAtomic Bk = 97

ToAtomic Cf = 98

ToAtomic Es = 99

ToAtomic Fm = 100

ToAtomic Md = 101

ToAtomic No = 102

ToAtomic Lr = 103

ToAtomic Rf = 104

ToAtomic Db = 105

ToAtomic Sg = 106

ToAtomic Bh = 107

ToAtomic Hs = 108

ToAtomic Mt = 109

ToAtomic Ds = 110

ToAtomic Rg = 111

ToAtomic Cn = 112

ToAtomic Nh = 113

ToAtomic Fl = 114

ToAtomic Mc = 115

ToAtomic Lv = 116

ToAtomic Ts = 117

ToAtomic Og = 118