Reactions

Description

Documentation

data Hydrocarbon

newtype wrapper for a Hydrocarbon Molecule

data MolReaction

Data type representing some reaction with reactants and products. Termed "inRxn" and "outRxn" to leave "reactants" and "products" as future getters when reactions are validated on creation.

Constructors

MolReaction

Fields

inRxn :: [Molecules]

outRxn :: [Molecules]

asMolecule :: Molecular a => a -> Molecule

Yields Molecule given a wrapped Molecular compound

type family ValidHydrocarbon (cCt :: Nat) (hCt :: Nat) :: Nat where ...

Type family yielding the order for single- (-ane), double- (-ene), or triple- (-yne) bonded hydrocarbon Emits a TypeError for none of the above

Equations

ValidHydrocarbon cCt hCt = If (hCt == ((2 * cCt) + 2)) 1 (If (hCt == (2 * cCt)) 2 (If (hCt == ((2 * cCt) - 2)) 3 (TypeError (((Text "Invalid simple Hydrocarbon: C" :<>: ShowType cCt) :<>: Text "H") :<>: ShowType hCt))))

mkHydrocarbon :: forall cCt hCt bondClass m. (KnownNat cCt, KnownNat hCt, bondClass ~ ValidHydrocarbon cCt hCt, m ~ If (cCt == 1) (Central 0 Neutral C '[Terminating 1 Neutral H hCt]) (Central 0 Neutral C (Terminating 1 Neutral H (4 - bondClass) ': '[Replicate4H (Central bondClass Neutral C '[]) (cCt - 2)])), KnownMolecule m) => Hydrocarbon

Reifies a Hydrocarbon given a number of Cs and Hs

type CO2 = Central 0 Neutral C '[Terminating 2 Neutral O 2]

Helpful type synonym for the Molecule representing Carbon Dioxide

type H2O = Central 0 Neutral O '[Terminating 1 Neutral H 2]

Helpful type synonym for the Molecule representing Water

type O2 = Terminating 0 Neutral O 2

Helpful type synonym for the Molecule representing Oxygen

combustHydrocarbon :: Hydrocarbon -> MolReaction

Yields the MolReaction representing the complete combustion of some Hydrocarbon

deltaHBondEnergy :: Floating a => MolReaction -> a

Uses bondEnergy to compute approximate ethalpy change for a gaseous reaction Relatively inaccurate at present