Slaters

Description

Documentation

screenConst :: forall a. Fractional a => EConf -> a

Yields screening constant based on a value-level Electron Configuration using Slater's Rules

zEff :: forall a. Fractional a => Element -> a

Yields effective nuclear charge (for valence e-) Atomic number - screening constant

zEffIon :: forall a st. (KnownCharge st, Fractional a) => Species st -> Maybe a

Yields effective nuclear charge for an ion Atomic number - ion EConf screening constant

rCov :: forall a. Floating a => a -> a

Conversion from max atomic radius to covalent radius Approximate using regression on available elements

eNegAR :: Fractional a => a -> a -> a

Allred-Rochow Electronegativity using zEff and rCov

eNeg :: Floating a => a -> a -> a

Pauling electronegativity based on regression from eNegAR

percentIonic :: Floating a => a -> a -> a

Percent Ionic Character using Pauling's formula for electronegativity difference