Slaters.Safe

Description

Warning: There is little benefit to a type-level interface given all elements through Oganesson are considered 'Valid.' Subject to removal unless further functionality neccessitates a type-level interface. All functions are extremely slow past Sodium, since they calculate the full electron configuration

Documentation

type ValidPQN (n :: Nat) = (1 <= n, n <= 7, KnownNat n)

Constraint representing valid n for which there is an n*

effPQN :: forall n a. (ValidPQN n, Fractional a) => a

Reifies some n as its associated n*

orbitalExp :: forall e a n. (Fractional a, KnownElem e, ValidPQN n, n ~ ANum2PQN (ToAtomic e)) => a

Yields an orbital exponent, (Z-S)/n*, given some type-level Element satisfying ValidPQN

rMax :: forall e a n. (Fractional a, KnownElem e, ValidPQN n, n ~ ANum2PQN (ToAtomic e)) => a

Reifies some Element satisfying ValidPQN as its max atomic radius in pm (53 pma0)*(n orbitalExp)

rCov' :: forall e a n. (Floating a, KnownElem e, ValidPQN n, n ~ ANum2PQN (ToAtomic e)) => a

Approximate covalent radius for some type-level Element

eNegAR' :: forall e a n. (Floating a, KnownElem e, ValidPQN n, n ~ ANum2PQN (ToAtomic e)) => a

Allred-Rochow electronegativity given some type-level Element

eNeg' :: forall e a n. (Floating a, KnownElem e, ValidPQN n, n ~ ANum2PQN (ToAtomic e)) => a

Approximate Pauling electronegativity given some type-level Element

percentIonic' :: forall e1 e2 a n1 n2. (Floating a, KnownElem e1, KnownElem e2, ValidPQN n1, n1 ~ ANum2PQN (ToAtomic e1), ValidPQN n2, n2 ~ ANum2PQN (ToAtomic e2)) => a

Percent Ionic character using Pauling electronegativity for two type-level Elements