Compounds

Documentation

data Ionic (e1Ct :: Nat) (e2Ct :: Nat)

Basic Ionic compound made up of two Species and associated counts

Constructors

forall a b.(KnownNat a, KnownNat b, '(e1Ct, e2Ct) ~ BinCoeffs a b) => Ionic (Species (Pos a)) (Species (Neg b))

data Covalent2 (e1Ct :: Nat) (e2Ct :: Nat)

Shortcut to a generic binary covalent compound Incompatible with Molecule as of now due to struggle forming a Lewis Structure Subject to removal following better Lewis Structure generation

mkPolarCov2 :: forall e1 e1Ct e2 e2Ct. (KnownElem e1, KnownElem e2, ValidPolar2 e1 e1Ct e2 e2Ct Neutral) => Covalent2 e1Ct e2Ct

Yields a Covalent2 from two elements and associated counts satisfying ValidPolar2

mkMaybeIonic :: forall e1 e1Ct e2 e2Ct c1 c2. (KnownElem e1, KnownElem e2, ValidPolar2 e1 e1Ct e2 e2Ct Neutral, ValidIonic e1 e1Ct e2 e2Ct c1 c2) => Either (Covalent2 e1Ct e2Ct) (Ionic e1Ct e2Ct)

Reifies a binary compound as Ionic if it has at least 50 percentIonic` character, otherwise Covalent2

type ValidIonic (e1 :: Element) (e1Ct :: Nat) (e2 :: Element) (e2Ct :: Nat) (c1 :: Nat) (c2 :: Nat) = (Species (Pos c1) ~ IonRep e1, Species (Neg c2) ~ IonRep e2, KnownNat e1Ct, KnownNat e2Ct, KnownNat c1, KnownNat c2, ValidIonic_ e1Ct e2Ct (BinCoeffs c1 c2), '(e1Ct, e2Ct) ~ BinCoeffs c1 c2)

Constraint representing a valid formula unit of two representative Elements in their common ions

data Molecule

Tree-like Molecule, composed of central and terminating elements Position, Bonds to parent, Formal Charge, Element, Children|Count

Constructors

Central Nat Charge Element [Molecule]

Terminating Nat Charge Element Nat

data Molecules

Equivalent of Atoms for Molecule

Constructors

Molecules Molecule Int

class Molecular a where

Useful for functions on newtype-wrapped Molecules satisfying special constraints

Methods

asMolecule :: a -> Molecule

type KnownMolecule (m :: Molecule) = (KnownMolecPart m, FC0 (FormalCharge m))

Representation of a reifiable Molecule with a total formal charge of zero

molecValI :: forall m. KnownMolecule m => Molecule

Reifies a type-level Molecule independent of Proxy

shatterMolecule :: Molecule -> [Atoms]

Breaks apart a Molecule into the Atoms it is composed of in its current structure

shatterSimple :: Molecule -> [Atoms]

shatterMolecule in whicch the structure is cleaned until only one set of Atoms for each Element exists