ElectronConf

Description

Documentation

data QL

Type representing states of the azimuthal or angular momentum quantum number, l

Constructors

SL

PL

DL

FL

class KnownQL (a :: QL)

Type class representing a reifiable azimuthal quantum number

Minimal complete definition

qlSing

qlVal :: forall q. KnownQL q => Proxy q -> QL

Reification of QL using Proxy

qlValI :: forall q. KnownQL q => QL

Reification of QL independent of Proxy

data Sublevel

Type representing a given sublevel For example, 1s2 = `SubL 1 SL 2`

Constructors

SubL Nat QL Nat

data EConf

Definition of electron configuration equivalent to a singly linked list

Constructors

Nucleus

Sublevel :<-: EConf infixr 5

type family GetN (ec :: Sublevel) :: Nat where ...

Type-level equivalent of getN

Equations

GetN (SubL n _ _) = n

peakN :: EConf -> Either EConf Sublevel

Should a DL or FL sublevel lower the PQN, returns the final SL sublevel

lastSubL :: EConf -> Either EConf Sublevel

Unwraps an EConf to Nucleus or the final Sublevel

anumToEConf :: Int -> EConf

Value-level equivalent of ANumToEConf

anumToSublevel :: Int -> Sublevel

anumToEConf yielding just the final Sublevel Value equivalent of ANumLastSL

type family ANum2PQN (z :: Nat) :: Nat where ...

Shortcut type-level conversion of atomic number to PQN

Equations

ANum2PQN 1 = 1

ANum2PQN 2 = 1

ANum2PQN 3 = 2

ANum2PQN 4 = 2

ANum2PQN z = GetN (ANumLastSL z (SubL 2 SL 2))

type family ANumLastSL (z :: Nat) (ec :: Sublevel) :: Sublevel where ...

Warning: Greatly lengthens type-checking, avoid if possible

Type-level conversion of atomic number to final sublevel, equivalent to case ANumToEConf z of sl :ec - sl Without extra expense

Equations

ANumLastSL 0 st = st

ANumLastSL z (SubL n SL e) = If (e == 1) (ANumLastSL (z - 1) (SubL n SL 2)) (If (n <=? 3) (ANumLastSL (z - 1) (SubL n PL 1)) (If (n <=? 5) (ANumLastSL (z - 1) (SubL (n - 1) DL 1)) (ANumLastSL (z - 1) (SubL (n - 2) FL 1))))

ANumLastSL z (SubL n PL e) = If (e <=? 5) (ANumLastSL (z - 1) (SubL n PL (e + 1))) (ANumLastSL (z - 1) (SubL (n + 1) SL 1))

ANumLastSL z (SubL n DL e) = If (e <=? 9) (ANumLastSL (z - 1) (SubL n DL (e + 1))) (ANumLastSL (z - 1) (SubL (n + 1) PL 1))

ANumLastSL z (SubL n FL e) = If (e <=? 13) (ANumLastSL (z - 1) (SubL n FL (e + 1))) (ANumLastSL (z - 1) (SubL (n + 1) DL 1))

type family ANumToEConf (n :: Nat) :: EConf where ...

Warning: Greatly lengthens type-checking, avoid if possible

Type level conversion of an atomic number to electron configuration Far too slow for elements past Sodium

Equations

ANumToEConf 0 = TypeError (Text "No Atomic # 0")

ANumToEConf 1 = SubL 1 SL 1 :<-: Nucleus

ANumToEConf 2 = SubL 1 SL 2 :<-: Nucleus

ANumToEConf 3 = SubL 2 SL 1 :<-: (SubL 1 SL 2 :<-: Nucleus)

ANumToEConf 4 = SubL 2 SL 2 :<-: (SubL 1 SL 2 :<-: Nucleus)

ANumToEConf a = AN2ECHelper (a - 4) (ANumToEConf 4)

addElectrons :: Int -> EConf -> EConf

Adding electrons to electron configuration Useful for anion formation or stepping through elements

addElectronsSL :: Int -> Sublevel -> Sublevel

addElectrons yielding just the final Sublevel

dropElectrons :: Int -> EConf -> EConf

Dropping electrons from an electron configuration Useful for cation formation or stepping through elements Yields no warning regarding being left with just the nucleus

dropElectronsSL :: Int -> Sublevel -> Either EConf Sublevel

dropElectrons yielding just the final Sublevel Left indicates just Nucleus

formIon :: Int -> EConf -> EConf

Formation of an ion electron configuration from Int charge and electron configuration

formIonSL :: Int -> Sublevel -> Either EConf Sublevel

Formation of an ion yielding just the final Sublevel Any Left value indicates just Nucleus

getN :: Integral a => Sublevel -> a

Extract Principal Quantum Number (n) from Sublevel

getMS :: Sublevel -> QMS

Extract Spin Quantum Number (ms) from Sublevel

getML :: Sublevel -> Int

Extract Magnetic Quantum Number (ml) from Sublevel

getL :: Sublevel -> QL

Extract Angular Momentum Quantum Number (l) from Sublevel

foldrEC :: (Sublevel -> b -> b) -> b -> EConf -> b

foldr definition as EConf cannot have a Foldable instance

slValI :: forall ec n l e. (KnownQL l, KnownNat n, KnownNat e, SubL n l e ~ ec) => Sublevel

Reification of a Sublevel independent of Proxy